Add analysis module (copied from Yann's PR branch)
