+++ fairgraph/__init__.py
@@ -23,3 +23,7 @@ import re
 from .client import KGClient
 
 __version__ = "0.5.2.dev"
+
+
+from . import (analysis, base, brainsimulation, client, commons, core, data, electrophysiology,
+               errors, minds, software, uniminds, utility)
