+++ fairgraph/brainsimulation.py
@@ -376,91 +376,6 @@ class EModel(ModelInstance):
         KGObject.__init__(self, **args)
 
 
-class AnalysisResult(KGObject):
-    """The result of a data analysis.
-
-    For example a graph, a histogram, etc. The result is expected to be stored either in a local
-    file or in a web-accessible location with a direct URL.
-
-    Note that local results files smaller than 1 MB in size will be uploaded and stored within
-    the Knowledge Graph. Larger files must be stored elsewhere.
-    """
-    namespace = DEFAULT_NAMESPACE
-    _path = "/simulation/analysisresult/v1.0.0"
-    type = ["prov:Entity", "nsg:Entity", "nsg:AnalysisResult"]
-    context = [
-        "{{base}}/contexts/neurosciencegraph/core/data/v0.3.1",
-        "{{base}}/contexts/nexus/core/resource/v0.3.0",
-        {
-            "wasDerivedFrom": "prov:wasDerivedFrom",
-            "generatedAtTime": "prov:generatedAtTime",
-            "wasAttributedTo": "prov:wasAttributedTo",
-            #"wasGeneratedBy": "prov:wasGeneratedBy",
-        }
-    ]
-    fields = (
-        Field("name", basestring, "name", required=True),
-        Field("result_file", (Distribution, basestring), "distribution"),
-        Field("timestamp", datetime, "generatedAtTime", default=datetime.now),
-        Field("derived_from", KGObject, "wasDerivedFrom"),
-        Field("attributed_to", Person, "wasAttributedTo"),
-        #Field("generated_by", Analysis, "wasGeneratedBy"),
-        Field("description", basestring, "description")
-    )
-    existence_query_fields = ("name", "timestamp")
-
-    def __init__(self, name, result_file=None, timestamp=None, derived_from=None,
-                 attributed_to=None, #generated_by=None,
-                 description=None, id=None, instance=None):
-        super(AnalysisResult, self).__init__(
-            name=name, result_file=result_file, timestamp=timestamp, derived_from=derived_from,
-            attributed_to=attributed_to, #generated_by=generated_by,
-            description=description, id=id, instance=instance
-        )
-        self._file_to_upload = None
-        if isinstance(result_file, basestring):
-            if result_file.startswith("http"):
-                self.result_file = Distribution(location=result_file)
-            elif os.path.isfile(result_file):
-                self._file_to_upload = result_file
-                self.result_file = None
-        elif result_file is not None:
-            for rf in as_list(self.result_file):
-                assert isinstance(rf, Distribution)
-
-    def save(self, client):
-        super(AnalysisResult, self).save(client)
-        if self._file_to_upload:
-            self.upload_attachment(self._file_to_upload, client)
-
-    def upload_attachment(self, file_path, client):
-        assert os.path.isfile(file_path)
-        statinfo = os.stat(file_path)
-        if statinfo.st_size > ATTACHMENT_SIZE_LIMIT:
-            raise Exception(
-                "File is too large to store directly in the KnowledgeGraph, please upload it to a Swift container")
-        # todo, use the Nexus HTTP client directly for the following
-        headers = client._nexus_client._http_client.auth_client.get_headers()
-        content_type, encoding = mimetypes.guess_type(file_path, strict=False)
-        response = requests.put("{}/attachment?rev={}".format(self.id, self.rev or 1),
-                                headers=headers,
-                                files={
-                                    "file": (os.path.basename(file_path),
-                                             open(file_path, "rb"),
-                                             content_type)
-        })
-        if response.status_code < 300:
-            logger.info("Added attachment {} to {}".format(file_path, self.id))
-            self._file_to_upload = None
-            self.result_file = Distribution.from_jsonld(response.json()["distribution"][0])
-        else:
-            raise Exception(str(response.content))
-
-    def download(self, local_directory, client):
-        for rf in as_list(self.result_file):
-            rf.download(local_directory, client)
-
-
 class ValidationTestDefinition(KGObject, HasAliasMixin):
     """Definition of a model validation test.
