+++ test/test_brainsimulation.py
@@ -10,7 +10,7 @@ from pyxus.resources.entity import Instance
 from fairgraph.base import KGQuery, KGProxy, as_list, Distribution, KGObject
 from fairgraph.commons import BrainRegion, CellType, QuantitativeValue
 from fairgraph.brainsimulation import (
-    ModelScript, ModelProject, ModelInstance, MEModel, Morphology, EModel, AnalysisResult,
+    ModelScript, ModelProject, ModelInstance, MEModel, Morphology, EModel,
     ValidationTestDefinition, ValidationScript, ValidationResult, ValidationActivity
 )
 from fairgraph.core import Person, use_namespace as use_core_namespace
@@ -22,7 +22,6 @@ use_core_namespace("modelvalidation")
 
 
 test_data_lookup.update({
-    "/v0/data/modelvalidation/simulation/analysisresult/v1.0.0/": "test/test_data/nexus/brainsimulation/analysisresult_list_0_10.json",
     "/v0/data/modelvalidation/simulation/emodel/v0.1.1/": "test/test_data/nexus/brainsimulation/emodel_list_0_10.json",
     "/v0/data/modelvalidation/simulation/memodel/v0.1.2/": "test/test_data/nexus/brainsimulation/memodel_list_0_10.json",
     "/v0/data/modelvalidation/simulation/modelinstance/v0.1.1/": "test/test_data/nexus/brainsimulation/modelinstance_list_0_10.json",
@@ -165,35 +164,6 @@ class TestEModel(BaseTestKG):
         cls = self.class_under_test
         objects = cls.list(kg_client, api="nexus", size=10)
         assert len(objects) == 10, len(objects)
-
-
-class TestAnalysisResult(BaseTestKG):
-    class_under_test = AnalysisResult
-
-    def test_list_nexus(self, kg_client):
-        cls = self.class_under_test
-        objects = cls.list(kg_client, api="nexus", size=10)
-        assert len(objects) == 10, len(objects)
-
-    def test_existence_query(self, kg_client):
-        obj = AnalysisResult("foo", timestamp=datetime(2000, 1, 1))
-        expected = {
-            "op": "and",
-            "value": [
-                {
-                    "path": "schema:name",
-                    "op": "eq",
-                    "value": "foo"
-                },
-                {
-                    "path": "prov:generatedAtTime",
-                    "op": "eq",
-                    "value": "2000-01-01T00:00:00"
-                }
-            ]
-        }
-        generated = obj._build_existence_query(api="nexus")
-        assert expected == generated
 
 
 class TestValidationTestDefinition(BaseTestKG):
